+++ pkgs/cupertino_http/ios/Classes/native_cupertino_bindings.m
@@ -1,1 +1,0 @@
-#include "../../src/native_cupertino_bindings.m"