+++ pkgs/flutter_http_example/lib/http_client_factory.dart
@@ -6,6 +6,7 @@ import 'dart:io';
 
 import 'package:cronet_http/cronet_http.dart';
 import 'package:cupertino_http/cupertino_http.dart';
+import 'package:flutter/widgets.dart';
 import 'package:http/http.dart';
 import 'package:http/io_client.dart';
 
@@ -13,6 +14,7 @@ const _maxCacheSize = 2 * 1024 * 1024;
 
 Client httpClient() {
   if (Platform.isAndroid) {
+    WidgetsFlutterBinding.ensureInitialized();
     final engine = CronetEngine.build(
         cacheMode: CacheMode.memory,
         cacheMaxSize: _maxCacheSize,