+++ pkgs/http/lib/src/client.dart
@@ -54,17 +54,17 @@ abstract interface class Client {
   /// Sends an HTTP POST request with the given headers and body to the given
   /// URL.
   ///
-  /// [body] sets the body of the request. It can be a [String], a [List<int>]
-  /// or a [Map<String, String>].
+  /// [body] sets the body of the request. It can be a `String`, a `List<int>`
+  /// or a `Map<String, String>`.
   ///
-  /// If [body] is a String, it's encoded using [encoding] and used as the body
-  /// of the request. The content-type of the request will default to
+  /// If [body] is a `String`, it's encoded using [encoding] and used as the
+  /// body of the request. The content-type of the request will default to
   /// "text/plain".
   ///
-  /// If [body] is a List, it's used as a list of bytes for the body of the
+  /// If [body] is a `List`, it's used as a list of bytes for the body of the
   /// request.
   ///
-  /// If [body] is a Map, it's encoded as form fields using [encoding]. The
+  /// If [body] is a `Map`, it's encoded as form fields using [encoding]. The
   /// content-type of the request will be set to
   /// `"application/x-www-form-urlencoded"`; this cannot be overridden.
   ///
@@ -77,15 +77,15 @@ abstract interface class Client {
   /// Sends an HTTP PUT request with the given headers and body to the given
   /// URL.
   ///
-  /// [body] sets the body of the request. It can be a [String], a [List<int>]
-  /// or a [Map<String, String>]. If it's a String, it's encoded using
+  /// [body] sets the body of the request. It can be a `String`, a `List<int>`
+  /// or a `Map<String, String>`. If it's a `String`, it's encoded using
   /// [encoding] and used as the body of the request. The content-type of the
   /// request will default to "text/plain".
   ///
-  /// If [body] is a List, it's used as a list of bytes for the body of the
+  /// If [body] is a `List`, it's used as a list of bytes for the body of the
   /// request.
   ///
-  /// If [body] is a Map, it's encoded as form fields using [encoding]. The
+  /// If [body] is a `Map`, it's encoded as form fields using [encoding]. The
   /// content-type of the request will be set to
   /// `"application/x-www-form-urlencoded"`; this cannot be overridden.
   ///
@@ -98,15 +98,15 @@ abstract interface class Client {
   /// Sends an HTTP PATCH request with the given headers and body to the given
   /// URL.
   ///
-  /// [body] sets the body of the request. It can be a [String], a [List<int>]
-  /// or a [Map<String, String>]. If it's a String, it's encoded using
+  /// [body] sets the body of the request. It can be a `String`, a `List<int>`
+  /// or a `Map<String, String>`. If it's a `String`, it's encoded using
   /// [encoding] and used as the body of the request. The content-type of the
   /// request will default to "text/plain".
   ///
-  /// If [body] is a List, it's used as a list of bytes for the body of the
+  /// If [body] is a `List`, it's used as a list of bytes for the body of the
   /// request.
   ///
-  /// If [body] is a Map, it's encoded as form fields using [encoding]. The
+  /// If [body] is a `Map`, it's encoded as form fields using [encoding]. The
   /// content-type of the request will be set to
   /// `"application/x-www-form-urlencoded"`; this cannot be overridden.
   ///
@@ -148,8 +148,11 @@ abstract interface class Client {
 
   /// Closes the client and cleans up any resources associated with it.
   ///
-  /// It's important to close each client when it's done being used; failing to
-  /// do so can cause the Dart process to hang.
+  /// Some clients maintain a pool of network connections that will not be
+  /// disconnected until the client is closed. This may cause programs using
+  /// using the Dart SDK (`dart run`, `dart test`, `dart compile`, etc.) to
+  /// not terminate until the client is closed. Programs run using the Flutter
+  /// SDK can still terminate even with an active connection pool.
   ///
   /// Once [close] is called, no other methods should be called. If [close] is
   /// called while other asynchronous methods are running, the behavior is
@@ -198,34 +201,30 @@ Client? get zoneClient {
 /// and the convenience HTTP functions (e.g. [http.get]). If [clientFactory]
 /// returns `Client()` then the default [Client] is used.
 ///
-/// When used in the context of Flutter, [runWithClient] should be called before
-/// [`WidgetsFlutterBinding.ensureInitialized`](https://api.flutter.dev/flutter/widgets/WidgetsFlutterBinding/ensureInitialized.html)
-/// because Flutter runs in whatever [Zone] was current at the time that the
-/// bindings were initialized.
-///
-/// [`runApp`](https://api.flutter.dev/flutter/widgets/runApp.html) calls
-/// [`WidgetsFlutterBinding.ensureInitialized`](https://api.flutter.dev/flutter/widgets/WidgetsFlutterBinding/ensureInitialized.html)
-/// so the easiest approach is to call that in [body]:
-/// ```
-/// void main() {
-///  var clientFactory = Client.new; // Constructs the default client.
-///  if (Platform.isAndroid) {
-///     clientFactory = MyAndroidHttpClient.new;
-///  }
-///  runWithClient(() => runApp(const MyApp()), clientFactory);
-/// }
-/// ```
+/// Only fresh `Client` instances using the default constructor are impacted.
+/// HTTP requests made using `dart:io` or `dart:html` APIs,
+/// or using specifically instantiated client implementations, are not affected.
 ///
 /// If [runWithClient] is used and the environment defines
 /// `no_default_http_client=true` then generated binaries may be smaller e.g.
 /// ```shell
-/// $ flutter build appbundle --dart-define=no_default_http_client=true ...
 /// $ dart compile exe --define=no_default_http_client=true ...
 /// ```
 ///
 /// If `no_default_http_client=true` is set then any call to the [Client]
 /// factory (i.e. `Client()`) outside of the [Zone] created by [runWithClient]
 /// will throw [StateError].
+///
+/// > [!IMPORTANT]
+/// > Flutter does not guarantee that callbacks are executed in a particular
+/// > [Zone].
+/// >
+/// > Instead of using [runWithClient], Flutter developers can use a framework,
+/// > such as [`package:provider`](https://pub.dev/packages/provider), to make
+/// > a [Client] available throughout their applications.
+/// >
+/// > See the
+/// > [Flutter Http Example](https://github.com/dart-lang/http/tree/master/pkgs/flutter_http_example).
 R runWithClient<R>(R Function() body, Client Function() clientFactory,
         {ZoneSpecification? zoneSpecification}) =>
     runZoned(body,