+++ pkgs/http_client_conformance_tests/lib/http_client_conformance_tests.dart
@@ -7,6 +7,7 @@ import 'package:http/http.dart';
 import 'src/close_tests.dart';
 import 'src/compressed_response_body_tests.dart';
 import 'src/isolate_test.dart';
+import 'src/multipart_tests.dart';
 import 'src/multiple_clients_tests.dart';
 import 'src/redirect_tests.dart';
 import 'src/request_body_streamed_tests.dart';
@@ -25,6 +26,7 @@ export 'src/close_tests.dart' show testClose;
 export 'src/compressed_response_body_tests.dart'
     show testCompressedResponseBody;
 export 'src/isolate_test.dart' show testIsolate;
+export 'src/multipart_tests.dart' show testMultipartRequests;
 export 'src/multiple_clients_tests.dart' show testMultipleClients;
 export 'src/redirect_tests.dart' show testRedirect;
 export 'src/request_body_streamed_tests.dart' show testRequestBodyStreamed;
@@ -64,6 +66,12 @@ export 'src/server_errors_test.dart' show testServerErrors;
 /// If [canReceiveSetCookieHeaders] is `false` then tests that require that
 /// "set-cookie" headers be received by the client will not be run.
 ///
+/// If [supportsFoldedHeaders] is `false` then the tests that assume that the
+/// [Client] can parse folded headers will be skipped.
+///
+/// If [supportsMultipartRequest] is `false` then tests that assume that
+/// multipart requests can be sent will be skipped.
+///
 /// The tests are run against a series of HTTP servers that are started by the
 /// tests. If the tests are run in the browser, then the test servers are
 /// started in another process. Otherwise, the test servers are run in-process.
@@ -74,8 +82,10 @@ void testAll(
   bool redirectAlwaysAllowed = false,
   bool canWorkInIsolates = true,
   bool preservesMethodCase = false,
+  bool supportsFoldedHeaders = true,
   bool canSendCookieHeaders = false,
   bool canReceiveSetCookieHeaders = false,
+  bool supportsMultipartRequest = true,
 }) {
   testRequestBody(clientFactory());
   testRequestBodyStreamed(clientFactory(),
@@ -86,12 +96,15 @@ void testAll(
       canStreamResponseBody: canStreamResponseBody);
   testRequestHeaders(clientFactory());
   testRequestMethods(clientFactory(), preservesMethodCase: preservesMethodCase);
-  testResponseHeaders(clientFactory());
+  testResponseHeaders(clientFactory(),
+      supportsFoldedHeaders: supportsFoldedHeaders);
   testResponseStatusLine(clientFactory());
   testRedirect(clientFactory(), redirectAlwaysAllowed: redirectAlwaysAllowed);
   testServerErrors(clientFactory());
   testCompressedResponseBody(clientFactory());
   testMultipleClients(clientFactory);
+  testMultipartRequests(clientFactory(),
+      supportsMultipartRequest: supportsMultipartRequest);
   testClose(clientFactory);
   testIsolate(clientFactory, canWorkInIsolates: canWorkInIsolates);
   testRequestCookies(clientFactory(),