+++ pkgs/http/CHANGELOG.md
@@ -1,4 +1,10 @@
-## 1.2.2-wip
+## 1.2.3-wip
+
+* Fixed unintended HTML tags in doc comments. 
+
+## 1.2.2
+
+* Require package `web: '>=0.5.0 <2.0.0'`.
 
 ## 1.2.1